[ADD] moved UploadCSV related types to a types.d.ts file
